Fixed bugs in procrustes analysis of PDM Added saving of procrustes analysis' result to jpeg file Deleted image of unwanted size

+++ Code/procrustes_analysis.py
@@ -126,7 +126,7 @@ def procrustes_analysis(reference_shape, shape):
     temp_sh = temp_sh / scale
     aligned_shape = rotate(temp_sh, theta)
 
-    return aligned_shape
+    return np.int32(aligned_shape)
 
 
 def procrustes_distance(reference_shape, shape):
